requirements.txt
asgiref==3.10.0
Django==5.2.8
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
